refactor: move the components-pane plugin initialization to the plugin side
