+++ packages/plugin-components-pane/build.json
@@ -1,6 +1,14 @@
 {
   "plugins": [
     "build-plugin-component",
+    [
+      "@alilc/build-plugin-alt",
+      {
+        "type": "plugin",
+        "inject": true,
+        "openUrl": "https://lowcode-engine.cn/demo/index.html?debug"
+      }
+    ],
     "build-plugin-fusion",
     [
       "build-plugin-moment-locales",
