+++ packages/plugin-components-pane/build.plugin.js
@@ -4,6 +4,7 @@ module.exports = ({ onGetWebpackConfig }) => {
   onGetWebpackConfig('component-demo-web', (config) => {
     config.plugin('HtmlWebpackPlugin').tap((options) => {
       const [config] = options;
+      console.log("Config: ", config);
       return [
         {
           ...config,
