+++ packages/plugin-components-pane/src/pane/index.module.scss
